+++ client/src/pages/user/SignIn.tsx
@@ -1,0 +1,12 @@
+
+import Loading from "../../components/common/Loading";
+
+
+export interface SignIn {
+  page: "SignIn"
+}
+
+
+export function SignIn(state: SignIn) {
+  return <Loading />;
+}
