+++ client/src/state/User.ts
@@ -1,12 +1,7 @@
-import {dispatch, Dispatch} from "./Dispatch";
+import {Dispatch} from "../core/Dispatch";
 
 import {User as UserState} from "../components/user/User";
-
-
 import {UserService} from "../service/UserService";
-
-
-// var service = new UserService();
 
 
 
@@ -21,7 +16,7 @@ async function loadUser() {
   if (user != null) {
     var userState = signedInUser(user.email ?? user.name);
 
-    dispatch(Dispatch.State(state => ({
+    Dispatch.send(Dispatch.State(state => ({
       ...state,
       user: userState
     })));
@@ -30,7 +25,7 @@ async function loadUser() {
 
 async function removeUser() {
   await UserService.get().startSignOut();
-  dispatch(Dispatch.State(state => ({
+  Dispatch.send(Dispatch.State(state => ({
     ...state,
     user: defaultUser
   })));
@@ -56,63 +51,3 @@ export const User = {
   loadUser: loadUser,
   signedInUser: signedInUser
 }
-
-
-// import Oidc from "oidc-client";
-// import * as P from "./ProtoLens";
-
-
-// // export interface NoUser {
-// //   kind: "NoUser";
-// // }
-
-// // export interface AUser {
-// //   kind: "AUser";
-// //   id: string;
-// //   name: string;
-// // }
-
-// export type User = {
-//   oidc: Oidc.UserManager | null;
-// };
-
-
-// // export const NoUser = {
-// //   instance: {
-// //     kind: "NoUser" as const
-// //   },
-
-// //   lens: <T>(id: P.Prism<T, NoUser>) => id
-// // }
-
-
-// // export const AUser = {
-// //   instance: {
-// //     kind: "AUser" as const
-// //   },
-
-// //   lens: <T>(id: P.Prism<T, AUser>) => id
-// // }
-
-
-// // let guardNoUser = (a: User): a is NoUser => (a.kind == "NoUser");
-// // let guardAUser = (a: User): a is AUser => (a.kind == "AUser");
-
-// export const User = {
-//   instance: () => ({
-//     oidc: null
-//   })
-//   //   return {
-//   //     oidc: new Oidc.UserManager({
-//   //       "authority": "https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration",
-//   //       "client_id": "d8ee60f3-f059-4169-93b4-8faf1c32a9d8",
-//   //       "redirect_uri": "http://localhost:32015/"
-//   //     }),
-//   //     ...NoUser.instance
-//   //   };
-//   // },
-
-//   // lens: <T>(id: P.Lens<T, User>) => id
-//     // .union("noUser", guardNoUser, NoUser.lens)
-//     // .union("aUser", guardAUser, AUser.lens),
-// }
