--- client/src/state/login/Login.ts
@@ -1,5 +1,0 @@
-
-export interface Login {
-
-}
-
