+++ client/src/state/page/read/StoryOverview.ts
@@ -1,98 +1,34 @@
-import request from "../../../Request";
-import {Dispatch} from "../../Dispatch";
+import {request} from "../../../service/RequestService";
+import {Dispatch} from "../../../core/Dispatch";
+import type {PageDescriptor} from "../../../core/PageDescriptor";
 import type {State} from "../../State";
-import type {PageState} from "../../model/PageState";
-import type {PageDescriptor} from "../../model/PageDescriptor";
-import {Lens} from "../../model/PageState";
+import type {Page} from "../../Page";
+import {Pages, Lens} from "../../Pages";
 
 import type {Story} from "../../../model/read/Story";
 
-import {Loadable, StoryOverview as Component} from "../../../components/page/read/StoryOverview";
+import {Loadable, StoryOverview as Component} from "../../../pages/read/StoryOverview";
 export type StoryOverview = Component;
 
 
 
 async function onLoad(dispatch: Dispatch, state: State) {
 
-  // console.log("onLoad 1");
-
   var loadingState = Lens.StoryOverview.stories.state.get(state.page);
 
   if (loadingState == "unloaded") {
-  // console.log("onLoad 2");
     dispatch(Dispatch.Update(Lens.StoryOverview.stories.set({ state: "loading" })));
-
-    // console.log("onLoad 3");
-
-    // await new Promise(resolve => setTimeout(resolve, 500));
 
     var storiesResponse = await request<Array<Story>>(
       "GET", "api/v1/read/stories");
-    // console.log("onLoad 5");
 
     if (storiesResponse.isResult) {
       dispatch(Dispatch.Update(Lens.StoryOverview.stories.set({ state: "loaded", result: storiesResponse.message })));
-
     }
     else {
-      // console.log("got error.");
       dispatch(Dispatch.Update(Lens.StoryOverview.stories.set({ state: "failed", error: storiesResponse.message })));
     }
-
-// p => {
-//       if (p.page == "StoryOverview") {
-//         return { ...p, stories: { kind: "loading" }};
-//       }
-
-//       return null;
-//     }));
-
-
   }
-
-  // console.log("onLoad 4");
-  // if (Lens..get(state.page) == "StoryOverview") {
-
-  // var page = state.page;
-  // if (page.page != "StoryOverview") {
-  //   return;
-  // }
-
-  // if (page.stories.kind == "unloaded") {
-  //   dispatch(Dispatch.Update(p => {
-  //     if (p.page == "StoryOverview") {
-  //       return { ...p, stories: { kind: "loading" }};
-  //     }
-
-  //     return null;
-  //   }));
-
-
-
-  //   switch (storiesResponse.kind) {
-  //     case "error":
-  //       console.log("request had an error.");
-  //       return;
-  //     case "result":
-  //       console.log("request succeeded", storiesResponse);
-  //       let stories = storiesResponse.message;
-
-  //       dispatch(Dispatch.Update(p => {
-  //         if (p.page == "StoryOverview") {
-  //           return instance({ kind: "loaded", result: stories });
-  //         }
-  //         return null;
-  //       }));
-
-  //       return;
-  //   }
-  // }
-
-  //   if (storiesResponse.isError) {
-  //     return;
-  //   }
-
-  // }
 }
