--- client/src/ui/useAsyncEffect.ts
@@ -1,9 +1,0 @@
-import * as PreactHooks from "preact/hooks";
-
-export default function useAsyncEffect(effect: () => Promise<void>) {
-  PreactHooks.useEffect(() => {
-    (async () => {
-      await effect();
-    })();
-  });
-}
